Time: 3 ms (57.65%), Space: 7.3 MB (37.86%) - LeetHub

+++ 0397-integer-replacement/0397-integer-replacement.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    unordered_map<long long,int>mp;
-    long long solve(int n){
+    unordered_map<int,int>mp;
+    int solve(int n){
         if(n==1) return 0;
         if(mp.find(n)!=mp.end()) return mp[n];
         if(n%2==0) return mp[n]=1+solve(n/2);
